atlas-backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
torch==2.9.1
torchvision==0.24.1
opencv-python==4.12.0.88
pillow==12.0.0
numpy==2.2.6
